feat: refactor Luma event types and cache structure for improved clarity and organization
