+++ src/pages/index.astro
@@ -4,7 +4,7 @@ import Layout from '../layouts/Layout.astro';
 import SltStackImg from '../components/icons/SltStackImg.astro';
 import BookAppointmentButton from '../components/BookAppointmentButton.astro';
 import EventBanner from '../components/EventBanner.astro';
-import type { LumaEvent } from '../helpers/lumaEvents.helper';
+import type { LumaEvent } from '../types/luma';
 import lumaCache from '../data/luma-event-cache.json';
 
 const nextEvent = (lumaCache.event ?? null) as LumaEvent | null;
